slicer_getbuildinfo: Gracefully handle itemId already removed or invalid
This commit skips the the lookup of package key associated with
itemId that are either already removed or invalid.

+++ etc/slicer_getbuildinfo/__main__.py
@@ -178,11 +178,15 @@ def main():
                 packagesByItemId[itemId] = key
 
         with sqlite3.connect(dbfile) as db:
-            print(f"Removed {len(itemIdsToRemove)} rows")
+            print(f"Removing {len(itemIdsToRemove)} rows")
             for itemId in itemIdsToRemove:
+                if itemId not in packagesByItemId:
+                    print(f"  {itemId} (not found)")
+                    continue
                 db.execute("delete from _ where item_id=?", (itemId, ))
                 print(f"  {itemId} ({packagesByItemId[itemId]})")
             db.commit()
+            print(f"Removed {db.total_changes} rows")
 
         print("Saved {0}".format(dbfile))
